+++ packages/bluesky_cli/lib/src/command/commands.dart
@@ -13,7 +13,8 @@ import 'bsky_command.dart';
 import 'common/delete.dart';
 import 'feed/actor_feeds.dart';
 import 'feed/feed.dart';
-import 'feed/generator.dart';
+import 'feed/create_generator.dart';
+import 'feed/feed_generator.dart';
 import 'feed/like.dart';
 import 'feed/likes.dart';
 import 'feed/post.dart';
@@ -64,8 +65,9 @@ final feedCommands = <BskyCommand>[
   FeedCommand(),
   ThreadCommand(),
   RepostedByCommand(),
-  GeneratorCommand(),
+  CreateGeneratorCommand(),
   ActorFeedsCommand(),
+  FeedGeneratorCommand(),
 ];
 
 final notificationCommands = <BskyCommand>[