+++ packages/bluesky_cli/CHANGELOG.md
@@ -2,8 +2,9 @@
 
 ## v0.2.1
 
-- Supported `generator` command. ([#430](https://github.com/myConsciousness/atproto.dart/issues/430))
+- Supported `create-generator` command. ([#430](https://github.com/myConsciousness/atproto.dart/issues/430))
 - Supported `actor-feeds` command. ([#431](https://github.com/myConsciousness/atproto.dart/issues/431))
+- Supported `feed-generator` command. ([#433](https://github.com/myConsciousness/atproto.dart/issues/433))
 
 ## v0.2.0